Fix : 주문시 배송지 정보 연동

See merge request cloud_track/class_05/web_project2/team02/sdz!99

+++ src/main/java/com/elice/sdz/order/service/OrderService.java
@@ -13,9 +13,6 @@ import com.elice.sdz.order.entity.OrderDetail;
 import com.elice.sdz.order.repository.OrderDetailRepository;
 import com.elice.sdz.order.repository.OrderRepository;
 import com.elice.sdz.orderItem.dto.OrderItemDTO;
-import com.elice.sdz.orderItem.entity.OrderItem;
-import com.elice.sdz.orderItem.entity.OrderItemDetail;
-import com.elice.sdz.orderItem.repository.OrderItemRepository;
 import com.elice.sdz.product.entity.Product;
 import com.elice.sdz.product.repository.ProductRepository;
 import com.elice.sdz.user.entity.Users;
@@ -25,7 +22,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
@@ -89,21 +85,35 @@ public class OrderService {
             }
 
         }
-        // 배송 주소 생성 및 저장
-        DeliveryAddress deliveryAddress = new DeliveryAddress();
-        deliveryAddress.setUser(user);
-        deliveryAddress.setDeliveryAddressId(orderReqDto.getDeliveryAddress().getDeliveryAddressId());
-        deliveryAddress.setDeliveryAddress1(orderReqDto.getDeliveryAddress().getDeliveryAddress1());
-        deliveryAddress.setDeliveryAddress2(orderReqDto.getDeliveryAddress().getDeliveryAddress2());
-        deliveryAddress.setDeliveryAddress3(orderReqDto.getDeliveryAddress().getDeliveryAddress3());
-        deliveryAddress.setReceiverName(orderReqDto.getDeliveryAddress().getReceiverName());
-        deliveryAddress.setReceiverContact(orderReqDto.getDeliveryAddress().getReceiverContact());
-        deliveryAddress.setDeliveryRequest(orderReqDto.getDeliveryAddress().getDeliveryRequest());
-        deliveryAddress.setDefaultCheck(false);
 
+        DeliveryAddress deliveryAddress;
+        if (orderReqDto.isNewAddress()) {
+            deliveryAddress = new DeliveryAddress();
+            deliveryAddress.setUser(user);
+            deliveryAddress.setDeliveryAddressId(orderReqDto.getDeliveryAddress().getDeliveryAddressId());
+            deliveryAddress.setDeliveryAddress1(orderReqDto.getDeliveryAddress().getDeliveryAddress1());
+            deliveryAddress.setDeliveryAddress2(orderReqDto.getDeliveryAddress().getDeliveryAddress2());
+            deliveryAddress.setDeliveryAddress3(orderReqDto.getDeliveryAddress().getDeliveryAddress3());
+            deliveryAddress.setReceiverName(orderReqDto.getDeliveryAddress().getReceiverName());
+            deliveryAddress.setReceiverContact(orderReqDto.getDeliveryAddress().getReceiverContact());
+            deliveryAddress.setDeliveryRequest(orderReqDto.getDeliveryAddress().getDeliveryRequest());
+            deliveryAddress.setDefaultCheck(false);
 
-        DeliveryAddress savedDeliveryAddress = deliveryAddressRepository.save(deliveryAddress);
+            deliveryAddressRepository.save(deliveryAddress);
+        } else {
+            deliveryAddress = deliveryAddressRepository.findById(orderReqDto.getDeliveryAddress().getDeliveryAddressId())
+                    .orElseThrow(() -> new CustomException(ErrorCode.DELIVERY_ADDRESS_NOT_FOUND));
 
+            if (orderReqDto.isAddressModified()) {
+                deliveryAddress.setDeliveryAddress1(orderReqDto.getDeliveryAddress().getDeliveryAddress1());
+                deliveryAddress.setDeliveryAddress2(orderReqDto.getDeliveryAddress().getDeliveryAddress2());
+                deliveryAddress.setDeliveryAddress3(orderReqDto.getDeliveryAddress().getDeliveryAddress3());
+                deliveryAddress.setReceiverName(orderReqDto.getDeliveryAddress().getReceiverName());
+                deliveryAddress.setReceiverContact(orderReqDto.getDeliveryAddress().getReceiverContact());
+                deliveryAddress.setDeliveryRequest(orderReqDto.getDeliveryAddress().getDeliveryRequest());
+                deliveryAddressRepository.save(deliveryAddress);
+            }
+        }
 
         // 주문 객체 생성
         Order order = Order.builder()
@@ -118,7 +128,7 @@ public class OrderService {
         // 배송 정보 설정
         Delivery delivery = new Delivery();
         delivery.setOrder(order);
-        delivery.setDeliveryAddress(savedDeliveryAddress);
+        delivery.setDeliveryAddress(deliveryAddress);
         delivery.setDeliveryStatus(Delivery.Status.PENDING);
         order.setDelivery(delivery);
 
@@ -178,7 +188,6 @@ public class OrderService {
         return orderRepository.findAll().stream()
                 .map(this::toResDto)
                 .collect(Collectors.toList());
-
     }
 
     //관리자 주문 상태수정
@@ -191,23 +200,42 @@ public class OrderService {
     }
     //관리자 주문 삭제
     @Transactional
-    public void deleteOrderByAdmin(Long id) {
-        Order order = orderRepository.findById(id)
-                .orElseThrow(() -> new CustomException(ErrorCode.ORDER_NOT_FOUND));
-        // 주문에 연결된 OrderDetail 목록을 조회
-        List<OrderDetail> orderDetails = order.getOrderDetails();
-
-        // 각 주문 상품에 대해 재고 처리
-        for (OrderDetail detail : orderDetails) {
-            Product product = detail.getProduct();
-            if (product != null) {
-                product.setProductCount(product.getProductCount() + detail.getOrderCount());
-                productRepository.save(product);
-            }
+    public void deleteOrdersByAdmin(List<Long> orderIds) {
+        if (orderIds.isEmpty()) {
+            throw new CustomException(ErrorCode.ORDER_NOT_FOUND); // 삭제할 주문 ID가 없을 경우
         }
 
-        orderRepository.delete(order);
+        // 주문 ID에 해당하는 주문들을 조회
+        List<Order> orders = orderRepository.findAllById(orderIds);
+        if (orders.size() != orderIds.size()) {
+            throw new CustomException(ErrorCode.ORDER_NOT_FOUND); // 주문이 존재하지 않는 경우
+        }
+
+        // 각 주문에 대해 처리
+        for (Order order : orders) {
+            try {
+                // 주문에 연결된 OrderDetail 목록을 조회
+                List<OrderDetail> orderDetails = order.getOrderDetails();
+
+                // 주문에 연결된 각 상품에 대해 재고 처리
+                for (OrderDetail detail : orderDetails) {
+                    Product product = detail.getProduct();
+                    if (product != null) {
+                        product.setProductCount(product.getProductCount() + detail.getOrderCount());
+                        productRepository.save(product);
+                    }
+                }
+
+                // 주문 삭제
+                orderRepository.delete(order);
+            } catch (Exception e) {
+                log.error("주문 {} 삭제 중 오류 발생", order.getOrderId(), e);
+                throw new CustomException(ErrorCode.INTERNAL_SERVER_ERROR); // 예외 발생 시 처리
+            }
+        }
     }
+
+
     //Entity->toDto 변환
     private OrderResDto toResDto(Order order) {
         OrderResDto dto = new OrderResDto();
